Initial Content (Start of Project)
I added a homepage where it greets the user and has a welcoming section along with a features section, a contact section, and an about section. User can view homepage and then login by clicking login on the navbar where it redirects the user to the login page.

PLEASE EDIT OR CHANGE THE STYLE IF YOU WANT THIS IS JUST FOR US TO GET STARTED!