Updated project code
The navbar scripts were out of sync with each webpage. Each webpage is now running the same scripts for the navbar. The navbar now fully works on each page. Also added in code where when a user adds in a new semester and when the user clicks the newly created semester it launches the user to the classes page for that chosen semester.