Updates Classes page
Now when we press the add another classes button the current data in the input fields gets saved to the database. Users can choose more than one day a week since some classes are sometimes offered twice a week. The weekday is set to as a boolean value in the database.